Preliminary New Environment
The NewKitchen Scene is the beginning of what I intend to be the final environment for the game to take place in. I still need to place enemies, spawners, and bake a navmesh into them, but this is a start. May add more decoratives in the future.

I also gave the player a chefs hat.